feat(schema-engine-wasm): wasm-compatible `sql-schema-describer`, `schema-connector` (#5106)
* feat(schema-engine-wasm): wasm-compatible sql-schema-describer

* fix(schema-engine-wasm): wasm compilation, move "SqlSchemaDescriberBackend" to sqlite-native impl

* fix(schema-engine-wasm): tests

* feat(schema-engine-wasm): wasm-compatible schema-connector

* fix(schema-engine-wasm): tests

* fix(schema-engine-wasm): tests

* fix(sql-schema-describer): tests

* fix(schema-engine-wasm): Cargo.toml features of schema-connector and sql-schema-describer

* chore(quaint): workaround for mongodb-schema-connector

---------
