+++ quaint/src/lib.rs
@@ -104,6 +104,9 @@
 //! # }
 //! ```
 
+// TODO: remove once `quaint` is no longer a transitive dependency of `mongodb-schema-connector`.
+#![allow(dead_code)]
+
 #[cfg(not(any(feature = "sqlite", feature = "postgresql", feature = "mysql", feature = "mssql")))]
 compile_error!("one of 'sqlite', 'postgresql', 'mysql' or 'mssql' features must be enabled");
