+++ schema-engine/sql-migration-tests/Cargo.toml
@@ -7,7 +7,9 @@ edition = "2021"
 psl = { workspace = true, features = ["all"] }
 schema-core = { path = "../core" }
 sql-schema-connector = { path = "../connectors/sql-schema-connector" }
-sql-schema-describer = { path = "../sql-schema-describer" }
+sql-schema-describer = { path = "../sql-schema-describer", features = [
+    "all-native",
+] }
 user-facing-errors = { path = "../../libs/user-facing-errors", features = [
     "all-native",
 ] }
