+++ schema-engine/sql-schema-describer/src/sqlite.rs
@@ -2,18 +2,19 @@
 
 use crate::{
     getters::Getter, ids::*, parsers::Parser, Column, ColumnArity, ColumnType, ColumnTypeFamily, DefaultValue,
-    DescriberResult, ForeignKeyAction, Lazy, PrismaValue, Regex, SQLSortOrder, SqlMetadata, SqlSchema,
-    SqlSchemaDescriberBackend,
+    DescriberResult, ForeignKeyAction, Lazy, PrismaValue, Regex, SQLSortOrder, SqlSchema,
 };
 use either::Either;
 use indexmap::IndexMap;
 use quaint::{
     ast::{Value, ValueType},
-    connector::{ColumnType as QuaintColumnType, GetRow, ToColumnNames},
     prelude::ResultRow,
 };
 use std::{any::type_name, borrow::Cow, collections::BTreeMap, convert::TryInto, fmt::Debug, path::Path};
 use tracing::trace;
+
+#[cfg(feature = "sqlite-native")]
+pub(crate) mod native;
 
 #[async_trait::async_trait]
 pub trait Connection {
@@ -22,31 +23,6 @@ pub trait Connection {
         sql: &'a str,
         params: &'a [quaint::prelude::Value<'a>],
     ) -> quaint::Result<quaint::prelude::ResultSet>;
-}
-
-#[async_trait::async_trait]
-impl Connection for std::sync::Mutex<quaint::connector::rusqlite::Connection> {
-    async fn query_raw<'a>(
-        &'a self,
-        sql: &'a str,
-        params: &'a [quaint::prelude::Value<'a>],
-    ) -> quaint::Result<quaint::prelude::ResultSet> {
-        let conn = self.lock().unwrap();
-        let mut stmt = conn.prepare_cached(sql)?;
-        let column_types = stmt.columns().iter().map(QuaintColumnType::from).collect::<Vec<_>>();
-        let mut rows = stmt.query(quaint::connector::rusqlite::params_from_iter(params.iter()))?;
-        let column_names = rows.to_column_names();
-        let mut converted_rows = Vec::new();
-        while let Some(row) = rows.next()? {
-            converted_rows.push(row.get_result_row().unwrap());
-        }
-
-        Ok(quaint::prelude::ResultSet::new(
-            column_names,
-            column_types,
-            converted_rows,
-        ))
-    }
 }
 
 #[async_trait::async_trait]
@@ -67,32 +43,6 @@ pub struct SqlSchemaDescriber<'a> {
 impl Debug for SqlSchemaDescriber<'_> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct(type_name::<SqlSchemaDescriber<'_>>()).finish()
-    }
-}
-
-#[async_trait::async_trait]
-impl SqlSchemaDescriberBackend for SqlSchemaDescriber<'_> {
-    async fn list_databases(&self) -> DescriberResult<Vec<String>> {
-        Ok(self.get_databases().await?)
-    }
-
-    async fn get_metadata(&self, _schema: &str) -> DescriberResult<SqlMetadata> {
-        let mut sql_schema = SqlSchema::default();
-        let table_count = self.get_table_names(&mut sql_schema).await?.len();
-        let size_in_bytes = self.get_size().await?;
-
-        Ok(SqlMetadata {
-            table_count,
-            size_in_bytes,
-        })
-    }
-
-    async fn describe(&self, _schemas: &[&str]) -> DescriberResult<SqlSchema> {
-        self.describe_impl().await
-    }
-
-    async fn version(&self) -> DescriberResult<Option<String>> {
-        Ok(Some(quaint::connector::sqlite_version().to_owned()))
     }
 }
